passing tests

+++ compiler/test_gen/src/gen_tags.rs
@@ -634,7 +634,7 @@ fn nested_tag_union() {
                 "#
         ),
         ((41, 0), 0),
-        ((i64, i64), i64)
+        ((i64, u8), u8)
     );
 }
 #[test]
@@ -799,14 +799,14 @@ fn alignment_in_multi_tag_construction() {
         indoc!(
             r"#
                 x : [ Three Bool I64, Empty ]
-                x = Three (1 == 1) 32
+                x = Three (1 == 1) 32 
 
                 x
 
                 #"
         ),
-        (32i64, true, 1),
-        (i64, bool, i64)
+        ((32i64, true), 1),
+        ((i64, bool), u8)
     );
 
     assert_evals_to!(
@@ -818,8 +818,8 @@ fn alignment_in_multi_tag_construction() {
                 x
                 #"
         ),
-        (32i64, true, 2u8, 1),
-        (i64, bool, u8, i64)
+        ((32i64, true, 2u8), 1),
+        ((i64, bool, u8), u8)
     );
 }
 
@@ -948,8 +948,8 @@ fn nested_recursive_literal() {
                 #"
         ),
         0,
-        &(i64, i64, i64),
-        |x: &(i64, i64, i64)| x.2
+        &(i64, i64, u8),
+        |x: &(i64, i64, u8)| x.2
     );
 }
